Create CrudRepository for Comments

+++ src/main/java/fi/tuni/tiko/bloggingsite/BlogPostRepository.java
@@ -2,5 +2,5 @@ package fi.tuni.tiko.bloggingsite;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BlogPostRepository extends CrudRepository<BlogPost, Integer> {
+public interface BlogPostRepository extends CrudRepository<BlogPost, Long> {
 }
